screen for cache and store data

+++ lib/scripts/ios_sh.sh
@@ -2,9 +2,13 @@
 
 if [ "$#" -eq 1 ]; then
     IOS_PATH=`xcrun simctl get_app_container booted com.wrench.app2 data`
-    cp $IOS_PATH/Documents/$1.hive ./lib/scripts/
+    if [ -f $IOS_PATH/Documents/$1.hive ]; then
+        cp $IOS_PATH/Documents/$1.hive ./lib/scripts/
+    else
+        echo "Invalid BoxName"
+    fi
 else
-    echo "Argument missing."
+    echo "HiveBoxName is missing"
 fi
